requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.3
python-multipart==0.0.9
authlib==1.2.1
httpx==0.26.0
itsdangerous==2.1.2
python-slugify==8.0.1
pyotp==2.9.0
qrcode==7.4.2
fastapi-mail==1.4.1
pydantic[email]==2.6.1
python-dotenv==1.0.0
